requirements/base.txt
requests>=2.28
python-decouple~=3.6